refactored and updated tests
