+++ src/extensions/ContentDiff/helpers/index.js
@@ -1,9 +1,4 @@
-import createAssetHtml from './createAssetHtml';
-import getValue from './getValue';
-import getArrayValue, { arrayListTestId, arrayListItemTestId } from './getArrayValue';
-
+export * from './simpleObjects';
 export * from './lookups';
 export * from './createHtml';
 export * from './getters';
-
-export { getValue, createAssetHtml, getArrayValue, arrayListTestId, arrayListItemTestId };
